Update PortalHooks.php
- reverted previous edit.

That does not work with either condition. Somehow those are not set at the time that paritcular hook is used.

We can leave it as it is a small file and minimal variables being filled. Otherwise I need to use a different hook that has those arrays filled with the necessary data.

Signed-off-by: Chen Zhen <github.underdog@gmail.com>